Add GitHub Action workflow to run tests on each PR and commit
This is a small workflow to run automated tests so that it is possible to see from the PR page whether tests are passing or not.

There is a node_modules cache that is invalidated whenever package.json changes. It would be better to use package-lock.json or better yet npm-shrinkwrap.json but that is out of scope.